fix: refactor code
- move main to pokeGo
- create helper function
- rename InterfaceRepository into GlobalInterface
- check for each layer connection

+++ internal/usecase/player.go
@@ -1,13 +1,15 @@
 package usecase
 
 import (
+	"errors"
 	"pokemon/internal/domain"
 	"pokemon/internal/interfaces"
 	"pokemon/internal/repository"
+	"pokemon/internal/util"
 )
 
 type PlayerUseCaseInterface interface {
-	interfaces.InterfaceRepository[domain.Player]
+	interfaces.GlobalInterface[domain.Player]
 }
 
 type PlayerUseCase struct {
@@ -20,25 +22,42 @@ func NewPlayerUseCase(repo repository.PlayerRepositoryInterface) PlayerUseCaseIn
 			playerRepo: rp,
 		}
 	}
-	return nil
+	return PlayerUseCase{
+		playerRepo: nil,
+	}
 }
 
 func (uc PlayerUseCase) GetAll() (players []domain.Player, err error) {
+	if !util.IsConnected(uc.playerRepo) {
+		return nil, errors.New("player repository is not initialized")
+	}
 	return uc.playerRepo.GetAll()
 }
 
 func (uc PlayerUseCase) GetById(id int) (Player domain.Player, err error) {
+	if !util.IsConnected(uc.playerRepo) {
+		return domain.Player{}, errors.New("player repository is not initialized")
+	}
 	return uc.playerRepo.GetById(id)
 }
 
 func (uc PlayerUseCase) Save(Player *domain.Player) (err error) {
+	if !util.IsConnected(uc.playerRepo) {
+		return errors.New("player repository is not initialized")
+	}
 	return uc.playerRepo.Save(Player)
 }
 
 func (uc PlayerUseCase) UpdateById(Player *domain.Player) error {
+	if !util.IsConnected(uc.playerRepo) {
+		return errors.New("player repository is not initialized")
+	}
 	return uc.playerRepo.UpdateById(Player)
 }
 
 func (uc PlayerUseCase) DeleteById(id int) error {
+	if !util.IsConnected(uc.playerRepo) {
+		return errors.New("player repository is not initialized")
+	}
 	return uc.playerRepo.DeleteById(id)
 }
