+++ internal/usecase/pokemon.go
@@ -1,13 +1,15 @@
 package usecase
 
 import (
+	"errors"
 	"pokemon/internal/domain"
 	"pokemon/internal/interfaces"
 	"pokemon/internal/repository"
+	"pokemon/internal/util"
 )
 
 type PokemonUseCaseInterface interface {
-	interfaces.InterfaceRepository[domain.Pokemon]
+	interfaces.GlobalInterface[domain.Pokemon]
 }
 
 type PokemonUseCase struct {
@@ -20,25 +22,42 @@ func NewPokemonUseCase(repo repository.PokemonRepositoryInterface) PokemonUseCas
 			pokemonRepo: rp,
 		}
 	}
-	return PokemonUseCase{}
+	return PokemonUseCase{
+		pokemonRepo: nil,
+	}
 }
 
 func (uc PokemonUseCase) GetAll() (pokemons []domain.Pokemon, err error) {
+	if !util.IsConnected(uc.pokemonRepo) {
+		return nil, errors.New("pokemon repository is not initialized")
+	}
 	return uc.pokemonRepo.GetAll()
 }
 
 func (uc PokemonUseCase) GetById(id int) (pokemon domain.Pokemon, err error) {
+	if !util.IsConnected(uc.pokemonRepo) {
+		return domain.Pokemon{}, errors.New("pokemon repository is not initialized")
+	}
 	return uc.pokemonRepo.GetById(id)
 }
 
 func (uc PokemonUseCase) Save(pokemon *domain.Pokemon) (err error) {
+	if !util.IsConnected(uc.pokemonRepo) {
+		return errors.New("pokemon repository is not initialized")
+	}
 	return uc.pokemonRepo.Save(pokemon)
 }
 
 func (uc PokemonUseCase) UpdateById(pokemon *domain.Pokemon) error {
+	if !util.IsConnected(uc.pokemonRepo) {
+		return errors.New("pokemon repository is not initialized")
+	}
 	return uc.pokemonRepo.UpdateById(pokemon)
 }
 
 func (uc PokemonUseCase) DeleteById(id int) error {
+	if !util.IsConnected(uc.pokemonRepo) {
+		return errors.New("pokemon repository is not initialized")
+	}
 	return uc.pokemonRepo.DeleteById(id)
 }
